+++ src/Psl/Exception/RuntimeException.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Psl\Exception;
 
-class RuntimeException extends \RuntimeException implements ExceptionInterface
+use RuntimeException as RuntimeRootException;
+
+class RuntimeException extends RuntimeRootException implements ExceptionInterface
 {
 }