+++ src/Psl/Str/Byte/slice.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Psl\Str\Byte;
 
-use Psl;
+use Psl\Str;
 
 use function substr;
 
@@ -19,11 +19,11 @@ use function substr;
  *
  * @pure
  *
- * @throws Psl\Exception\InvariantViolationException If $offset is out-of-bounds.
+ * @throws Str\Exception\OutOfBoundsException If $offset is out-of-bounds.
  */
 function slice(string $string, int $offset, ?int $length = null): string
 {
-    $offset = Psl\Internal\validate_offset($offset, length($string));
+    $offset = Str\Internal\validate_offset($offset, length($string));
 
     return null === $length ? substr($string, $offset) : substr($string, $offset, $length);
 }