+++ src/Psl/Str/Grapheme/contains_ci.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Psl\Str\Grapheme;
 
 use Psl;
+use Psl\Str;
 
 /**
  * Returns whether the 'haystack' string contains the 'needle' string.
@@ -16,12 +17,13 @@ use Psl;
  *
  * @pure
  *
- * @throws Psl\Exception\InvariantViolationException If the $offset is out-of-bounds.
+ * @throws Str\Exception\OutOfBoundsException If $offset is out-of-bounds.
+ * @throws Psl\Exception\InvariantViolationException If unable to split $string into grapheme clusters.
  */
 function contains_ci(string $haystack, string $needle, int $offset = 0): bool
 {
     if ('' === $needle) {
-        return Psl\Internal\validate_offset($offset, length($haystack), true);
+        return Str\Internal\validate_offset($offset, length($haystack), true);
     }
 
     return null !== search_ci($haystack, $needle, $offset);