+++ src/Psl/Str/Grapheme/starts_with.php
@@ -4,10 +4,14 @@ declare(strict_types=1);
 
 namespace Psl\Str\Grapheme;
 
+use Psl;
+
 /**
  * Returns whether the string starts with the given prefix.
  *
  * @pure
+ *
+ * @throws Psl\Exception\InvariantViolationException If unable to split $string into grapheme clusters.
  */
 function starts_with(string $string, string $prefix): bool
 {