+++ src/Psl/Str/before_ci.php
@@ -4,19 +4,13 @@ declare(strict_types=1);
 
 namespace Psl\Str;
 
-use Psl;
-
 /**
- * @throws Psl\Exception\InvariantViolationException If the $offset is out-of-bounds.
+ * @throws Exception\OutOfBoundsException If the $offset is out-of-bounds.
  *
  * @pure
  */
-function before_ci(
-    string $haystack,
-    string $needle,
-    int $offset = 0,
-    Encoding $encoding = Encoding::UTF_8
-): ?string {
+function before_ci(string $haystack, string $needle, int $offset = 0, Encoding $encoding = Encoding::UTF_8): ?string
+{
     $length = search_ci($haystack, $needle, $offset, $encoding);
     if (null === $length) {
         return null;